# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/addStudent": {
            "post": {
                "description": "Creates a student with given data",
                "produces": [
                    "application/json"
                ],
                "summary": "Creates a student with given data",
                "operationId": "create-student",
                "parameters": [
                    {
                        "description": "student",
                        "name": "stud",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/src.Student"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/deleteStudent": {
            "delete": {
                "description": "Deletes a student with given id",
                "produces": [
                    "application/json"
                ],
                "summary": "Deletes a student with given id",
                "operationId": "delete-student",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/src.Student"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/getStudent": {
            "get": {
                "description": "Get the student with given id",
                "produces": [
                    "application/json"
                ],
                "summary": "Get the student with given id",
                "operationId": "get-student",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/src.Student"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/getStudents": {
            "get": {
                "description": "Get all students",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "students"
                ],
                "summary": "Get all students",
                "operationId": "get-students",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/src.Student"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "src.Student": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "rollNo": {
                    "type": "integer"
                }
            }
        }
    }
}